gcode-agent-cli/requirements.txt
google-generativeai>=0.5.0
python-dotenv>=1.0.0

# Dependencies for MCP server
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
mcp>=1.6.0 # Using version from search results

# Development/Testing
pytest>=7.0.0
pytest-mock>=3.0.0
